+++ src/interfaces/ClientSdkInstance.ts
@@ -5,6 +5,10 @@ import ApplyCouponPrices from './ApplyCouponPrices'
 
 import MessageType from '../enums/MessageType'
 import { ResignFormConfig, ResignRequest } from './ResignConfig'
+import {
+  WalletCardTypeEventName,
+  WalletCardTypeSubscriber
+} from './WalletCardType'
 
 export default interface ClientSdkInstance {
   init(config: InitConfig): Promise<void>
@@ -13,11 +17,12 @@ export default interface ClientSdkInstance {
     config?: ResignFormConfig
   ): Promise<ClientSdkInstance>
   update(config: UpdateConfig): Promise<void>
+  on(event: WalletCardTypeEventName, subscriber: WalletCardTypeSubscriber): void
   on<T extends MessageType>(
     event: T,
     callback: (e: MessageEvent<SdkMessage[T]>) => void
   ): void
-  unsubscribe(messageType: MessageType): void
+  unsubscribe(messageType: MessageType | WalletCardTypeEventName): void
   unsubscribeAll(): void
   submit(): void
   applyCoupon(couponCode: string): Promise<ApplyCouponPrices>